also check if message may contain a GPG signature
In some emails, the "multipart/pgp-signed" can be attached in the
main payload (which has a Content-Type "multipart/mixed").
So, the first check agianst a pgp-signed message fails.

Therefore, if the message is multipart, Ask (via GMail's API)
if the given message contains the words:
 BEGIN-PGP-SIGNATURE and END-PGP-SIGNATURE
and then (if so) get and verify that message.